add yums/offer

+++ database/migrations/2023_04_13_020600_create_clients_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('clients', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->tinyInteger('yums')->default(0);
+            $table->integer('yums')->default(0);
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
